Exercise 4
Write a C program to duplicate functions of a Bank ATM. Create a variable to holds
balance of your account and assign an initial value at the beginning. The ATM should
provide the following functionalities.
i) Display a welcome message
 Welcome to the ABC Bank!
ii) Then it will be display options to be selected.
Select one of the options from following list and enter
the number
1. Deposit 2. Withdraw 3. Check Balance
Your choice:
iii) If deposit is selected, display a message to enter the deposit amount.
Enter amount to deposit:
Then add the amount that user entered to the balance.
iv) if withdraw is selected, display a message to enter the withdrawal amount.
Enter amount to withdraw:
Then user input amount should be validated by checking whether the balance is
sufficient. If balance is less than the withdrawal amount a message should be displayed.
Your balance is Insufficient to withdraw that amount of
money
v) if Balance is selected, Display the balance.
 Your balance is 40000