Exercise 5
1. Write a C program to get a positive integer as user input and calculate the
sum of the numbers in the range from 0 to that number.
Enter a positive integer: 10
The Sum of range 0 - 10 is: 55
2. Write a C program to get a positive integer as user input and calculate the
factorial of the number. You should check the user input is whether
negative or not. If it’s negative, print a message “Factorial of a negative
number doesn't exist.”
Example:
Enter a positive integer: 10
The factorial of 10 is: 3628800
3. Write a C program to get a positive integer as user input and check whether
it is a prime number.
Example:
Enter a positive integer: 19
It is a prime number!