Exercise 6
1. Write the Bank ATM C program in Lab Exercise 6 using at least three functions
(deposit, withdraw and balanceCheck).
Write a C program to duplicate functions of a Bank ATM. Create a variable to hold the
balance of your account and assign an initial value at the beginning. The ATM should
provide the following functionalities.
i) Display a welcome message
 Welcome to the ABC Bank!
ii) Then it will be display options to be selected.
Select one of the options from following list and enter
the number
1. Deposit 2. Withdraw 3. Check Balance
Your choice:
iii) If deposit is selected, display a message to enter the deposit amount.
Enter amount to deposit:
Then add the amount that user entered to the balance.
iv) if withdraw is selected, display a message to enter the withdrawal amount.
Enter amount to withdraw:
Then user input amount should be validated by checking whether the balance is sufficient.
If the balance is less than the withdrawal amount a message should be displayed.
Your balance is Insufficient to withdraw that amount of
money
University of Sri Jayewardenepura, Sri Lanka
v) if Balance is selected, Display the balance.
 Your balance is 40000
2.
i) Write a C program to get the temperature in Celsius as a user input and
Convert it to Fahrenheit. The conversion should be done in a separate
function.
ii) Add another function to convert Fahrenheit to Celsius.
iii) Extend the program to get user choice, Fahrenheit or Celsius (similar to Bank
ATM) before the temperature is entered. Depending on the user’s choice, the
temperature is converted to a relevant unit using one of the functions created in i)
and ii).
3. a) Write a C program to create a simple calculator. First, it will prompt the user to
choose the operation choice (from 1 to 5). Then it asks the user to input two
integer values for the calculation. Each operation should be written in different
functions.
This program should run an infinite number of times until you stop the
execution.
See the sample below.
MENU
1. Add
2. Subtract
3. Multiply
4. Divide
5. Modulus
Enter your choice: 1
Enter your two numbers: 12 15
Result: 27