Exercise 7
1.
a) Write a C program to find the Greatest Common Divisor (GCD) of two (02) integer
numbers using recursion. The program should take two integers from the user and
display the GCD.
See the sample below.
Enter your two integer numbers: 8 12
The Greatest Common Divisor (GCD) of two numbers is: 4
Note:
The Greatest Common Divisor (GCD) of two or more integers is the largest positive
integer that divides each of the integers.
For example, GCD of 8 and 12 is 4
 gcd(8,12) = 4
b) Extend the above C program to find the Least Common Multiple (LCM) of two(2)
numbers. The program should take two integers from the user and display the LCM.
See the sample below.
Enter your two integer numbers: 6 10
The Least Common Multiple (LCM) of two numbers is: 30
Note:
The Least Common Multiple ( LCM ) is also referred to as the Lowest Common
Multiple ( LCM ) and Least Common Divisor ( LCD).
University of Sri Jayewardenepura, Sri Lanka
For two integers a and b, denoted LCM(a,b), the LCM is the smallest positive integer
that is evenly divisible by both a and b.
For example,
LCM(2,3) = 6
LCM(6,10) = 30